% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_scalefactor.R
\name{cal_scalefactor}
\alias{cal_scalefactor}
\title{cal_scalefactor}
\usage{
cal_scalefactor(z_score, percent_cut = 0.01)
}
\arguments{
\item{z_score}{a numeric vector that is original Z scores for individual cells calculated from gchromVAR}

\item{percent_cut}{a numeric number (0, 1) indicating what percentage of most enriched cells will be selected to calculate scale factor}
}
\value{
a numeric value
}
\description{
used for generate a scale factor from bias-corrected Z score for TRS scale and normalization
}
\examples{
\dontrun{ cal_scalefactor <- seedindx(z_score_mat$z_score, 0.01)}

}
