% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capOutlierQuantile.R
\name{capOutlierQuantile}
\alias{capOutlierQuantile}
\title{capOutlierQuantile}
\usage{
capOutlierQuantile(x, q_ceiling = 0.99)
}
\arguments{
\item{x}{a numeric vector}

\item{q_ceiling}{a numeric value indicating which quantile is used
for ceiling the vector.}
}
\value{
numeric vector
}
\description{
cap/normalize the trs with quantile
}
\examples{
x_ceiling <- capOutlierQuantile(x=seq_len(10), q_ceiling=0.8)
}
