% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_lsi.R
\name{do_lsi}
\alias{do_lsi}
\title{do_lsi}
\usage{
do_lsi(mat, dims = 30)
}
\arguments{
\item{mat}{(sparse matrix)  A saparse matrix of TF-IDF matrix (LSI) used for SVD}

\item{dims}{number of LSIs to calculate (default:30)}
}
\value{
a sparse matrix of LSI
}
\description{
Get a LSI matrix (cell-by-LSI) from TF-IDF matrix from single cell profiles
}
\examples{
\dontrun{ lsi_mat <- do_lsi(tfidf_mat, dims=30)}

}
