% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data}
\alias{example_data}
\title{Example data}
\usage{
example_data(
  name = c("pbmc5k_SE.rda", "mono.PP001.bed"),
  return_path = FALSE,
  colidx = 5
)
}
\arguments{
\item{name}{Name of the data to return.}

\item{return_path}{Whether to return just the path to the file (\code{TRUE})
or the imported  R object itself \code{FALSE}.}

\item{colidx}{The column index of the score. This
assumes that the first three columns specific genomic 
coordinates. Default is 5 for the fifth column in the .bed files}
}
\description{
Example data of multiple data types.
}
\examples{
pbmc5krda <- example_data(name="pbmc5k_SE.rda")
}
