% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmutualknn.R
\name{getmutualknn}
\alias{getmutualknn}
\title{getmutualknn}
\usage{
getmutualknn(lsimat = lsimat, num_k = 30)
}
\arguments{
\item{lsimat}{A dense or sparse matrix of cell (rows) x LSI (cols)}

\item{num_k}{A numeric input (30 by default) for k used for graph construction}
}
\value{
A sparse matrix indicating adjacent matrix from single cells used for graph construction
}
\description{
Get a mutual kNN graph (M-kNN) from the latent space (LSI) of single cell profiles
}
\examples{
\dontrun{
mknn_graph <- getmutualknn(lsimat=LSI_mat, num_k=30)
}


}
