% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_scale.R
\name{max_min_scale}
\alias{max_min_scale}
\title{max_min_scale}
\usage{
max_min_scale(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
a function for max-min normalization for network propagation score
}
\examples{
\dontrun{
trs <- max_min_scale(ceiling_np)}

}
