% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalk_sparse.R
\name{randomWalk_sparse}
\alias{randomWalk_sparse}
\title{randomWalk_sparse}
\usage{
randomWalk_sparse(
  intM,
  queryGenes,
  gamma = 0.05,
  seedWeight = "NO",
  stationary_cutoff = 1e-05
)
}
\arguments{
\item{intM}{a sparse matrix indicating the adjacent matrix (m x m, where m is the cell number) of cell-to-cell network (M-kNN graph)}

\item{queryGenes}{a logical vector indicating seed cells (TRUE) and non-seed cells (FALSE) with length of m, where m is the cell number. The length and position are corresponding to intM}

\item{gamma}{a numeric value indicating the probability of node restart at each step of random walk}

\item{seedWeight}{a numeric vector indicating the weight assigned to each node. "NO" (by default) means considering weights are equal (no weight)}

\item{stationary_cutoff}{delta used for determine the stationary state between any two adjacent iterations}
}
\value{
a numeric vector of network propagation score with length of m, where m is the cell number
}
\description{
Calculate the network propagation score using a set of seed cells and cell-to-cell graph
}
\examples{
\dontrun{
randomWalk_sparse <- tfidf(intM, queryGenes, gamma=0.05, seedWeight="NO", stationary_cutoff=1e-5)}

}
