% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{tfidf}
\alias{tfidf}
\title{tfidf}
\usage{
tfidf(bmat, mat_binary = TRUE, TF = TRUE, log_TF = TRUE, scale_factor = 1e+05)
}
\arguments{
\item{bmat}{a sparse/dense matrix or dataframe indicating count matrix (peak-by-cell matrix)}

\item{mat_binary}{a logic value to indicate if input bmat is spare matrix or not. bmat will be converted into a sparse matrix.}

\item{TF}{a logic value to indicate if term frequency (TF) normalization is performed}

\item{log_TF}{a logic value to indicate if natural logarithm values of the TF are calculated and used}

\item{scale_factor}{a scale factor used to multiple the resulting TF-IDF matrix}
}
\value{
a sparse matrix of normalized TFIDF matrix
}
\description{
Get a normalized TFIDF matrix from peak-by-cell matrix from single cell profiles
}
\examples{
\dontrun{
tfidf_mat <- tfidf(bmat=peakbycellmatrix, mat_binary=TRUE, TF=TRUE, log_TF=TRUE)}

}
