% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_scalefactor.R
\name{cal_scalefactor}
\alias{cal_scalefactor}
\title{cal_scalefactor}
\usage{
cal_scalefactor(z_score, percent_cut = 0.01)
}
\arguments{
\item{z_score}{a numeric vector that is original Z scores for
individual cells calculated from gchromVAR.}

\item{percent_cut}{a numeric number (0, 1) indicating what percentage of
most enriched cells will be selected to calculate scale factor.}
}
\value{
a numeric value
}
\description{
Used for generate a scale factor from bias-corrected Z score
for TRS scale and normalization.
}
\examples{
z_score <- example_results$mono_mat$z_score
sf <- cal_scalefactor(z_score = z_score)
}
