% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_lsi.R
\name{do_lsi}
\alias{do_lsi}
\title{do_lsi}
\usage{
do_lsi(mat, dims = 30)
}
\arguments{
\item{mat}{(sparse matrix)  A saparse matrix of TF-IDF matrix (LSI) used for SVD}

\item{dims}{number of LSIs to calculate (default:30)}
}
\value{
a sparse matrix of LSI
}
\description{
Get a LSI matrix (cell-by-LSI) from TF-IDF matrix from single cell profiles
}
\examples{
SE_pbmc5k <- example_data(name="pbmc5k_SE.rda")
peak_by_cell_mat <- SummarizedExperiment::assay(SE_pbmc5k)
#### Downsample to speed up example ####
peak_by_cell_mat <- peak_by_cell_mat[seq_len(500), seq_len(100)]
tfidf_mat <- tfidf(bmat=peak_by_cell_mat)
lsi_mat <- do_lsi(mat=tfidf_mat, dims=30)
}
